feat: restricted runs allowed Supports a new top-level key in content filters: restricted_runs_allowed. The value will be a mapping of course key (str) to a list of course run key (str). Updates the customer-based contains_content_items API endpoint (`/api/v1/enterprise-customer/{enterprise_uuid}/contains_content_items/`) to serialize an additional key in the response body containing a mapping of restricted runs to catalog UUIDs for requested course keys. Fundamentally, this becomes a transposition of the reverse mapping encoded in the content filter. ENT-9358
