+++ docker-compose.yml
@@ -1,4 +1,3 @@
-version: "2.1"
 services:
   mysql:
     image: mysql:8.0.28-oracle
