+++ enterprise_catalog/apps/catalog/constants.py
@@ -108,6 +108,12 @@ PROGRAM_TYPES_MAP = {
 
 FORCE_INCLUSION_METADATA_TAG_KEY = 'enterprise_force_included'
 
+RESTRICTED_RUNS_ALLOWED_KEY = 'restricted_runs_allowed'
+
+AGGREGATION_KEY_PREFIX = 'course:'
+
+COURSE_RUN_KEY_PREFIX = 'course-v1:'
+
 
 def json_serialized_course_modes():
     """
